bugfix: single-seq RCLO is incorrect given "N"
In RCLO: 1) forward and reverse complemented sequences are in one index; 2) the
reverse complement of the i-th smallest sequence is the i-th sequence in the
FM-index. We can achieve this by "revcomp | sort | revcomp". The multi-seq
version does this right, but the single-seq version is actually doing reverse
sorting, which is wrong.